@@ -94,6 +94,9 @@
     ]
   },
   "dependencies": {
+    "@babel/core": "^7.3.4",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/register": "^7.0.0",
     "@elastic/datemath": "5.0.2",
     "@elastic/eui": "9.5.0",
     "@elastic/filesaver": "1.1.2",
@@ -123,10 +126,7 @@
     "angular-sanitize": "1.6.5",
     "angular-sortable-view": "0.0.15",
     "autoprefixer": "^9.1.0",
-    "babel-core": "6.26.3",
-    "babel-loader": "7.1.5",
-    "babel-polyfill": "6.26.0",
-    "babel-register": "6.26.0",
+    "babel-loader": "8.0.5",
     "bluebird": "3.5.3",
     "boom": "^7.2.0",
     "brace": "0.11.1",
@@ -263,7 +263,7 @@
     "@octokit/rest": "^15.10.0",
     "@types/angular": "1.6.50",
     "@types/angular-mocks": "^1.7.0",
-    "@types/babel-core": "^6.25.5",
+    "@types/babel__core": "^7.1.0",
     "@types/bluebird": "^3.1.1",
     "@types/boom": "^7.2.0",
     "@types/chance": "^1.0.0",
@@ -275,7 +275,7 @@
     "@types/delete-empty": "^2.0.0",
     "@types/elasticsearch": "^5.0.30",
     "@types/enzyme": "^3.1.12",
-    "@types/eslint": "^4.16.2",
+    "@types/eslint": "^4.16.6",
     "@types/execa": "^0.9.0",
     "@types/fetch-mock": "7.2.1",
     "@types/getopts": "^2.0.0",
@@ -320,8 +320,8 @@
     "@types/zen-observable": "^0.8.0",
     "angular-mocks": "1.4.7",
     "archiver": "^3.0.0",
-    "babel-eslint": "^9.0.0",
-    "babel-jest": "^23.6.0",
+    "babel-eslint": "^10.0.1",
+    "babel-jest": "^24.1.0",
     "backport": "4.4.1",
     "chai": "3.5.0",
     "chance": "1.0.10",
@@ -335,17 +335,17 @@
     "enzyme-adapter-react-16": "^1.9.0",
     "enzyme-adapter-utils": "^1.10.0",
     "enzyme-to-json": "^3.3.4",
-    "eslint": "^5.6.0",
-    "eslint-config-prettier": "^3.1.0",
-    "eslint-plugin-babel": "^5.2.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^21.26.2",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-mocha": "^5.2.0",
+    "eslint": "^5.15.1",
+    "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jest": "^22.3.0",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-mocha": "^5.3.0",
     "eslint-plugin-no-unsanitized": "^3.0.2",
     "eslint-plugin-prefer-object-spread": "^1.2.1",
-    "eslint-plugin-prettier": "^2.6.2",
-    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-react": "^7.12.4",
     "faker": "1.1.0",
     "fetch-mock": "7.3.0",
     "geckodriver": "1.12.2",
@@ -356,7 +356,7 @@
     "grunt-karma": "2.0.0",
     "grunt-peg": "^2.0.1",
     "grunt-run": "0.7.0",
-    "gulp-babel": "^7.0.1",
+    "gulp-babel": "^8.0.0",
     "gulp-sourcemaps": "2.6.4",
     "has-ansi": "^3.0.0",
     "image-diff": "1.6.0",
@@ -402,9 +402,6 @@
     "supertest": "^3.1.0",
     "supertest-as-promised": "^4.0.2",
     "tree-kill": "^1.1.0",
-    "ts-jest": "^23.1.4",
-    "ts-loader": "^5.2.2",
-    "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
     "tslint-config-prettier": "^1.15.0",
     "tslint-microsoft-contrib": "^6.0.0",
