@@ -17,16 +17,7 @@
  * under the License.
  */
 
-const mockCreateLayout = jest.fn();
-jest.mock('../layouts/layouts', () => {
-  const { schema } = require('@kbn/config-schema');
-  return {
-    Layouts: {
-      configSchema: schema.object({ kind: schema.literal('mock') }),
-      create: mockCreateLayout,
-    },
-  };
-});
+import { mockCreateLayout } from './appenders.test.mocks';
 
 import { LegacyAppender } from '../../legacy/logging/appenders/legacy_appender';
 import { Appenders } from './appenders';
