@@ -7,7 +7,7 @@ import { CallCluster, CallClusterWithRequest } from 'src/legacy/core_plugins/ela
 import { Request, Server } from 'src/legacy/server/kbn_server';
 import { SavedObjectsClient } from 'src/legacy/server/saved_objects';
 
-import moment = require('moment');
+import moment from 'moment';
 import { XPackInfo } from 'x-pack/plugins/xpack_main/server/lib/xpack_info';
 import { ReindexSavedObject, ReindexStatus } from '../../../common/types';
 import { CredentialStore } from './credential_store';
