@@ -19,20 +19,33 @@
 
 module.exports = {
   presets: [
-    require.resolve('babel-preset-react'),
+    require.resolve('@babel/preset-typescript'),
+    require.resolve('@babel/preset-react')
   ],
   plugins: [
     require.resolve('babel-plugin-add-module-exports'),
-    // stage 3
-    require.resolve('babel-plugin-transform-async-generator-functions'),
-    require.resolve('babel-plugin-transform-object-rest-spread'),
 
-    // the class properties proposal was merged with the private fields proposal
+    // The class properties proposal was merged with the private fields proposal
     // into the "class fields" proposal. Babel doesn't support this combined
     // proposal yet, which includes private field, so this transform is
     // TECHNICALLY stage 2, but for all intents and purposes it's stage 3
     //
     // See https://github.com/babel/proposals/issues/12 for progress
-    require.resolve('babel-plugin-transform-class-properties'),
+    require.resolve('@babel/plugin-proposal-class-properties'),
   ],
+  overrides: [
+    {
+      // Babel 7 don't support the namespace feature on typescript code.
+      // With namespaces only used for type declarations, we can securely
+      // strip them off for babel on x-pack infra plugin
+      //
+      // See https://github.com/babel/babel/issues/8244#issuecomment-466548733
+      test: /x-pack[\/\\]plugins[\/\\]infra[\/\\].*[\/\\]graphql/,
+      plugins: [
+        [
+          require.resolve('babel-plugin-typescript-strip-namespaces'),
+        ],
+      ]
+    }
+  ]
 };
