@@ -25,7 +25,7 @@ function babelRegister() {
   const plugin = pluginConfig();
 
   try {
-    // add support for moved babel-register source: https://github.com/elastic/kibana/pull/13973
+    // add support for moved @babel/register source: https://github.com/elastic/kibana/pull/13973
     require(resolve(plugin.kibanaRoot, 'src/setup_node_env/babel_register')); // eslint-disable-line import/no-dynamic-require
   } catch (error) {
     if (error.code === 'MODULE_NOT_FOUND') {
