@@ -19,7 +19,7 @@
 
 require('./main.scss');
 
-import 'babel-polyfill';
+import '@babel/polyfill';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
