@@ -17,17 +17,7 @@
  * under the License.
  */
 
-const mockReaddir = jest.fn();
-const mockReadFile = jest.fn();
-const mockStat = jest.fn();
-jest.mock('fs', () => ({
-  readdir: mockReaddir,
-  readFile: mockReadFile,
-  stat: mockStat,
-}));
-
-const mockPackage = new Proxy({ raw: {} as any }, { get: (obj, prop) => obj.raw[prop] });
-jest.mock('../../../../../package.json', () => mockPackage);
+import { mockPackage, mockReaddir, mockReadFile, mockStat } from './plugin_discovery.test.mocks';
 
 import { resolve } from 'path';
 import { BehaviorSubject } from 'rxjs';
