@@ -17,14 +17,7 @@
  * under the License.
  */
 
-const mockWriteFileAsync = jest.fn();
-const mockMakeDirAsync = jest.fn();
-jest.mock('./utils', () => ({
-  // Jest typings don't define `requireActual` for some reason.
-  ...(jest as any).requireActual('./utils'),
-  writeFileAsync: mockWriteFileAsync,
-  makeDirAsync: mockMakeDirAsync,
-}));
+import { mockMakeDirAsync, mockWriteFileAsync } from './integrate_locale_files.test.mocks';
 
 import path from 'path';
 import { integrateLocaleFiles, verifyMessages } from './integrate_locale_files';
