@@ -59,11 +59,6 @@ export default {
   coverageReporters: [
     'html',
   ],
-  globals: {
-    'ts-jest': {
-      skipBabel: true,
-    },
-  },
   moduleFileExtensions: [
     'js',
     'json',
@@ -83,8 +78,7 @@ export default {
     'integration_tests/'
   ],
   transform: {
-    '^.+\\.js$': '<rootDir>/src/dev/jest/babel_transform.js',
-    '^.+\\.tsx?$': '<rootDir>/src/dev/jest/ts_transform.js',
+    '^.+\\.(js|tsx?)$': '<rootDir>/src/dev/jest/babel_transform.js',
     '^.+\\.txt?$': 'jest-raw-loader',
     '^.+\\.html?$': 'jest-raw-loader',
   },
