@@ -18,6 +18,6 @@
  */
 
 // Note: In theory importing the polyfill should not be needed, as Babel should
-// include the necessary polyfills when using `babel-preset-env`, but for some
+// include the necessary polyfills when using `@babel/preset-env`, but for some
 // reason it did not work. See https://github.com/elastic/kibana/issues/14506
 import '../../../setup_node_env/babel_register/polyfill';
