@@ -20,11 +20,13 @@
 import { AggParam } from '../../../agg_types';
 import { AggConfig } from '../../agg_config';
 
-interface AggParamEditorProps<T> {
+// NOTE: we cannot export the interface with export { InterfaceName }
+// as there is currently a bug on babel typescript transform plugin for it
+// https://github.com/babel/babel/issues/7641
+//
+export interface AggParamEditorProps<T> {
   agg: AggConfig;
   aggParam: AggParam;
   value: T;
   setValue(value: T): void;
 }
-
-export { AggParamEditorProps };
