@@ -19,13 +19,13 @@
 
 var resolve = require('path').resolve;
 
-// this must happen before `require('babel-register')` and can't be changed
+// this must happen before `require('@babel/register')` and can't be changed
 // once the module has been loaded
 if (!process.env.BABEL_CACHE_PATH) {
   process.env.BABEL_CACHE_PATH = resolve(__dirname, '../../../optimize/.babelcache.json');
 }
 
-// paths that babel-register should ignore
+// paths that @babel/register should ignore
 var ignore = [
   /\/bower_components\//,
   /\/kbn-pm\/dist\//,
@@ -48,10 +48,10 @@ var ignore = [
 if (global.__BUILT_WITH_BABEL__) {
   // when building the Kibana source we replace the statement
   // `global.__BUILT_WITH_BABEL__` with the value `true` so that
-  // when babel-register is required for the first time by users
+  // when @babel/register is required for the first time by users
   // it will exclude kibana's `src` directory.
   //
-  // We still need babel-register for plugins though, we've been
+  // We still need @babel/register for plugins though, we've been
   // building their server code at require-time since version 4.2
   // TODO: the plugin install process could transpile plugin server code...
   ignore.push(resolve(__dirname, '../../../src'));
@@ -66,10 +66,11 @@ if (global.__BUILT_WITH_BABEL__) {
 
 // modifies all future calls to require() to automatically
 // compile the required source with babel
-require('babel-register')({
+require('@babel/register')({
   ignore,
   babelrc: false,
   presets: [
     require.resolve('@kbn/babel-preset/node_preset')
   ],
+  extensions: ['.js', '.ts', '.tsx'],
 });
