@@ -37,8 +37,7 @@ export function createJestConfig({
       '**/*.test.{js,ts,tsx}'
     ],
     transform: {
-      '^.+\\.js$': `${kibanaDirectory}/src/dev/jest/babel_transform.js`,
-      '^.+\\.tsx?$': `${kibanaDirectory}/src/dev/jest/ts_transform.js`,
+      '^.+\\.(js|tsx?)$': `${kibanaDirectory}/src/dev/jest/babel_transform.js`,
     },
     transformIgnorePatterns: [
       '[/\\\\]node_modules[/\\\\].+\\.js$'
