@@ -5,7 +5,7 @@
  */
 
 const { resolve } = require('path');
-const register = require('babel-register');
+const register = require('@babel/register');
 const options = {
   babelrc: false,
   presets: [require.resolve('@kbn/babel-preset/node_preset')],
