@@ -8,9 +8,8 @@ import { format as formatUrl } from 'url';
 
 import elasticsearch from 'elasticsearch';
 import shieldPlugin from '../../../../server/lib/esjs_shield_plugin';
-import { TestInvoker } from '../lib/types';
 
-export function EsProvider({ getService }: TestInvoker) {
+export function EsProvider({ getService }) {
   const config = getService('config');
 
   return new elasticsearch.Client({
